requirements.txt
nonebot2[fastapi]>=2.4.4
nonebot-adapter-onebot>=2.4.6
rapidocr-onnxruntime>=1.3.26
openai>=1.0.0
httpx>=0.28.1
